Clustering using 'Topic Modeling Tool'
It's was considered the columns below to create documents:

columns = ['1 - Title', '5 - What is the topic of the study?', '6 - What are the code attributes being captured?', '7 - What are the research questions of the study?', '8 - What is the empirical method used?', '9 - What were the targeted languages?', '11 - What are the treatments (independent variables)?', '13 - What attributes are being measured (dependent variables)?', '14 - How are they being measured?', '16 - What did the paper find out? What are the answers to the research questions?']

Considering 10 topics.